Minimise LoopState mutation on read
Currently, primarily due to timing logs, every read causes a loop state mutation in almost every process it touches (e.g. bookie, penciller and each related sst file for HEAD requests).

Loop state mutations require copies, at least of the size of a word per element (+2), assuming each element is a reference to box (and in a sense doesn't really change).  This is a small cost, but repeated for each operation, may have an impact on both CPU and GC.

This commit moves all timing logs related to requests to a new leveled_monitor process, so that countdowns and accumulators for timing logs aren't required on the Loopstate.  Where possible, LoopState changes on read are then minimised.

Outstanding potential issue is the SST fetch_cache.  It might be preferable to replace the sst local fetch_cache with a ledger_readcache (that gets flushed on every push_mem).  This would stop rotating SST loop state on fetch_cache hits.