Close in loop
If the send_after occurs immediately before the close, the prompted handle_cast may fail due to termination.

Now close will cast to reply after the cast, and any additional messages in the queue are processed.  A message received by handle_info in between will be ignored and not cast to prompt a log