Mas i370 patch e (#385)
Improvement to monitoring for efficiency and improved readability of logs and stats.

As part of this, where possible, tried to avoid updating loop state on READ messages in leveled processes (as was the case when tracking stats within each process). 

No performance benefits found with change, but improved stats has helped discover other potential gains.