Profiling and Refactoring
The initial testing of the `guess` method for object count estimation indicated it was surprisingly slow.  Counting 1:1000 keys took 10% of the time - although there is additional work necessary to skip using segment lists, this  intuitively appears to be out by an order of magnitude.

Profiling indicated that the majority of time was being spent in two places:

- within `expand_list_by_pointer/5`, in particular though the use of `++`;
- in the `find_pos/4` function, in particular repeated calls to `lists:member/2`.

the expand_list_by_pointers has been refactored and simplified.  The member check under find_pos can now be avoided where SegmentIds are not randomly distributed.

This has reduced the time to `guess` object counts by 65% in OTP 22 (and by 75% by also changing to OTP 26).

The outstanding problem, is that `find_pos` is still perhaps unexpectedly dominant in CPU time:

leveled_sst:'-segment_checker/1-fun-2-'/5                   400000000    19.31  15560374  [      0.04]

leveled_sst:find_pos/4                                      403085303    35.05  28242226  [      0.07]

Further, there appears to be a discrepancy between performance in eunit and ct conditions of this function.