Address BigList ++ SmallList
Within the SST path there was use of ++ or lists:append where potentially lists:append(BigList, SmallList) was being used when lists:append(SmallList, BigList) would be more efficient.

The fetching of trimmed blocks in queries is now change to not use a large fold function, and now uses blocks_required/7 - where further efficiency is made by guarding against deserialisation of blocks 1 and 5 if they are out of the range.

Most of the performance improvement is in the building of the hashlist.

The other refactoring causes some confusion due to the returning of revrsed lists abd binaryslot_trimmedlist has been renamed to binaryslot_trimmedreversed to reflect this.

There is a functional change in that calls to binaryslot_trimmedlist with a range and a SegCheckFun would not previously be trimmed - whereas they should be.