Mas d31 nhskv16sstpclv2 (#427)
* Change fullPerf

Change the fullPerf test to run more tests, but with fewer keys.

Given that RS of 512 is being pushed in Riak, 2M objects is till a 300M+ object cluster. 10M >> 1B.  so these are still reasonable sizes to test.

A profilePerf test also added to generate all the profiles base don 2M objects.

* Address BigList ++ SmallList

Within the SST path there was use of ++ or lists:append where potentially lists:append(BigList, SmallList) was being used when lists:append(SmallList, BigList) would be more efficient.

The fetching of trimmed blocks in queries is now change to not use a large fold function, and now uses blocks_required/7 - where further efficiency is made by guarding against deserialisation of blocks 1 and 5 if they are out of the range.

Most of the performance improvement is in the building of the hashlist.

The other refactoring causes some confusion due to the returning of revrsed lists abd binaryslot_trimmedlist has been renamed to binaryslot_trimmedreversed to reflect this.

There is a functional change in that calls to binaryslot_trimmedlist with a range and a SegCheckFun would not previously be trimmed - whereas they should be.

* Alter ++ order in binaryslot_reader

Normally SlotsTOPoint will be smaller, so prepend this to the fetched list.

This will allow for larger ITERATOR_SCANWIDTH at lower cost, potentially.

* Reduce duplication

Duplication between sst_getfilteredslots and sst_getfilteredrange.

Tries hard to avoid appending a short list to a long list.

Also ensures that through all paths results are trimmed to both the left and right.

* Extend test

Queries where previously all returning a large number of index entries - changes made to make number of entries per query more realistic.  Also an update process added to show difference between loading and rotating keys.

* Trim before accumulating

When deserialising blocks, trim them at that point prior to accumulation.

Standardise the functions used for trimming, and add a unit test to ensure that the accumulation handles empty blocks correctly.

* Relabel as AAE fold

* Review/update comments

* Reset default test

* Remove redundant define statements

* Increase timeout

The clerk may be doing a large fold - and 20s may not be sufficient.  Has been seen to fail to shut down at the end of perf_SUITE fullperf test.

* Test v5

Test mini-queries - where generally a small number of entries are returned

* Remove unnecessary case

Single statement works for all N > 1

* Revert to ctperf as default

Bump timeout for slower AWS instances