DJB Hash and integer overflow
The implementation of magic_hash was a straight interpretation of the DJB hash code.  Every byte of the key in turn is bxor'd against an accumulated hash (which is multiplied by 33 (i.e. bsl 5 plus itself) every loop.

In the original C code, due to the integer overflow of unsigned integers, every loop the extra bits beyond 4 bytes in the hash are automatically discarded.  However, in Erlang the integer type can be of an arbitrary size - so each loop carries a larger hash.  This means that with a 100-byte key (say), there would be a 500-bit integer hash by the end (which is then chopped back to 32 bits before returning).

This change explicitly restricts the hast to a 32-bit integer on every loop.  This doesn't change the result, but is more efficient (and this can be detected in the perf_SUITE profile test, especially when using.a larger bucket name to generate larger keys).