Avoid double-reverse and lists:last
The generate_binary_slot function is primarily called from within form_slot/7.  In this case form_slot/7 would have the slot (i.e. the list of keys to form the slot) in reverse, and so would reverse it before calling generate_binary_slot.  If the slot was a lookup type, the generate_binary_slot function would then re-reverse to be efficient on accumulate_positions.  It would also need the last key - which it would find through lists:last.

Now the generate_binary_slot function can receive the slot reversed, and so will only reverse once.  As it will always reverse once, it will always have the list in reverse and so can find the last key through the more efficient hd/1 function.

When forming level 0 there will now be an additional lists:reverse/1 call - but normally there are an order of magnitude more non-level-0 file creations for each level-0 creation ... so this is still a net gain.

The BuildTimings was originally used to try and profile this in production - which seems unnecessary given the ability to profile via perf_SUITE. This is also altered to reduce the situations where this is called, and reduce the number of required calls to os:timestamp/1 to make it.