Make key_dominates more efficient
the most common case is when the created file is not. abasement file and the key in list 2 is < key in list 1.  so avoid the separate function call to key_dominates_expanded in this situation.

The load profile prior to this change has:

leveled_sst:key_dominates/3                             16139172     1.60    820906  [      0.05]

leveled_sst:key_dominates_expanded/2                    16133866     3.62   1859403  [      0.12]

So for every key covered by every merge, there is a 0.05 microsecond cost on key_dominates/3, and and a 0.12 microsecond cost on key_dominates_expanded/2.

Following the change, the profile is:

leveled_sst:key_dominates_expanded/2                     1621644     0.20    103125  [      0.06]

leveled_sst:key_dominates/3                             16235173     3.22   1663231  [      0.10]

Now key_dominates_expanded/2 is called an order of magnitude less time, and key_dominates/3 is more efficient than the previous combined cost.