Modify filter
The Filter should expect to compare BETWEEN low AND high ... supporting high AND low may be confusing.

The filter should allow for identifier to be in any location in a comparator i.e.

string BETWEEN identifier AND identifier

integer > identifier

This make check some validity constraints harder at parse time (i.e. that High >= Low), but makes tse result more compatible.

For ease, of implementation IN can only be used for a  comparison between a string and a list of strings (with the identifier representing either part).