Binary split not string split
Using the binary match/split to break up binary strings is a faster than using string split, as pre-compiled patterns can be used (saving 0.16 microsec per application).

compiled patterns can be passed across nodes, so pre-compiling them in the parser makes the function non-specific.  So instead, compile on demand.  If, as expected, this process is to check many times - at least have it compile only the once.

Examining the eprof profile also indicated significant time in lists:sublist - so the process of updating the map has been simplified.  Empty attributes will not now be added - and in the case of regex, attributes are only added if all are produced as expected.