Skip forward on dropwhile/takewhile
If we're looking for a small range - it is probable that there will be a lot to drop before, and so skip forward.

With a block size of 56, will skip forward 12 if there's 24 left.

takewhile is relatively expensive as endkey_passed function not a straight comparator.  So try and skip over shorter ranges (12 and 6).

Profile test indicates an improvement on both mini_query and query.