SST sub-blocks and block types (#475)
Add block types, so that based on the content of blocks, a block can be split into different arrangements of sub-blocks.  This will allow for head/tail of the block with index entries to be exposed, to allow for more efficient 2i.  Also in general, when looking for the nth element, only a sub-block will need to be opened.

The code trades performance for neatness, with precise pattern matching used to avoid overheads with copying when concatenating lists.