Standardise formatting using erlfmt (#479)
* Standardise formatting using erlfmt

Adding check of format to github actions

* Exclude generated files from formatting

* Use correct filelib function

Previously checked a file and also not a directory - where is_regular/1 is specifically checking the same thing.

Avoids a coverage miss, now the line has been reformated (as in our tests it is always a file), the only variance is when it is a directory.

* Revise formatting

Exclude where formatting has created less readable code (e.g. leveled_sstblock).  Also where erlfmt splits across lines and isolates the fun keyworrd in a fun((_) -> _) definition - this can cause problems with some editors/linters.

* Resolve highlighting issue by removing parenthesis