Simplify and reduce the number of compares
form_slot is split between the no_lookup part (all form_slot loops start assuming no_lookup) and the lookup part (now in form_slot_lookup).  This reduces the branches required in each loop.

The first key check is now made at the end and only when ncessary.

The check for the ?IDX tag is now done explicitly rather than through another function call.

The key_dominates function has been simplified and reorganised - passing the reaping responsibility back up to form_slot, and reducing the branches and ensuring the first condition is the most common (K2 is the next key - as KVL1 is a small list merging into the bigger KVL2).

Frequently called functions are now smaller, and so are inlined.