Add shortcut in find_nextkeys/6
This loop is a major factor in the performance of any long-running query or fold.

Commonly, it would be normal to expect the "next key" to be most commonly found at the lower level.  This implements a shortcut, whereby when the next key is at the bottom level, there is no need to check all the other levels next time round.