Increase min_heap_size
For temporary processes that require a large heap - set a bigger heap at the start.  This covers the spawn process to check blocks in SST, and the a snapshot of the penciller.

Both of these are either very short-lived, and/or constrained in terms of concurrency.

Also SST files will start with a large heap, then revert back to default.

When splitting a number of slots, there is now a binary copy.  Each individual slot would otherwise be a reference to part of a bigger binary - but these slots are processed in turn.  If GC interrupts, then binary memory will not drop.